+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/pipeline/middleware/ResolveContext.kt
@@ -10,7 +10,7 @@ import com.github.blarc.gitlab.template.lint.plugin.notifications.sendNotificati
 import com.github.blarc.gitlab.template.lint.plugin.pipeline.Pass
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import com.github.blarc.gitlab.template.lint.plugin.settings.ProjectSettings
-import com.github.blarc.gitlab.template.lint.plugin.ui.settings.Remote
+import com.github.blarc.gitlab.template.lint.plugin.ui.settings.remote.Remote
 import com.github.blarc.gitlab.template.lint.plugin.widget.LintStatusEnum
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service